backend/requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Image processing
opencv-python-headless==4.9.0.80
Pillow==10.3.0
exifread==3.0.0
numpy==1.26.4

# ML — Vision (optional; app falls back to mock signals when unavailable)

# Geo
requests==2.31.0
h3==3.7.7
geopandas==0.14.4
shapely==2.0.4

# Templates & PDF
jinja2==3.1.4
weasyprint==62.3

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai==0.7.1
